Improve font lookup
